add afterResizeDblClick callback and jqGridAfterResizeDblClick event
New events will be called at the end of processing of the double-click
on the column resizer. It allows to make some custom actions after
autoResizing of columns for example.

See the following code as an example:

afterResizeDblClick: function (options) {
alert("column " + options.cmName + " is resized. The new column width is
" + options.cm.width);
}