Add options to hideCol/hideCol improve columnChooser performance
The changes:

1) the second parameter of `showHideCol` can be Boolean (`true` to show
and `false` to hide) and not only the string "" to show and the string
"none" to hide.

2) the methods `showHideCol`, `hideCol` and `showCol` support now
additional optional parameter, which can be helpful to improve the
performance of hiding/showing multiple columns. The last option
parameter could be object with the following Boolean properties:
`skipSetGroupHeaders`, `skipFeedback`, `skipSetGridWidth`. Specifying of
`true` values for any from the properties force skipping of the
corresponding calls inside of `showHideCol`.

3) The modification of `columnChooser` uses the new options of
`showCol`/`hideCol` to improve the performance. One can consider to use
`skipFeedback: true` later, but I didn't implemented it now because the
consequences of the changes needed be analysed carefully before.

Thanks @Wooya (https://github.com/Wooya) for his suggestion from [the
comment](https://github.com/free-jqgrid/jqGrid/issues/198#issuecomment-173470782),
which pushed me to make the changes.