use context option of jQuery.ajax. Remove $.jgrid.useJSON
Multiple calls of `jqGridImport` before pending request will be finished
could follow the usage of new options `o` inside of `complete` callback
of old Ajax requests. To fix the problem we save now the object `o` in
`context` parameter, which allows to use original (correct) options as
`this` inside  of `complete` callback.

Legacy `$.jgrid.useJSON` is removed from the usage.