preparing for publishing of free jqGrid 4.13.2
There are many small changes in free jqGrid code since the month after after publishing of 4.13.1. The publishing of new version 4.13.2 should place the latest code on CDNs () and simplify the usage of the most recent version of the code. Both CDNs should provides modules, which can be used for loading via RequireJs for example. Because of migration of the CDNs to HTTP/2 protocol it opens new possibilities. I remind that the loading of multiple small files via HTTP/2 protocol has minimal overhead over loading of one large file. It makes loading of separate modules much more interesting as before and reduced the total size of traffic (because of not all modules will be used typically).