add `column` property to options of createEl and bindEv inside of jqFilter
See https://github.com/free-jqgrid/jqGrid/issues/253 for the problem with *modifying* of `searchoptions` of the items of `columns` parameter of `jqFilter`. The `columns` parameter contains typically the deep **copy** of `colModel`. The new `column` property allows to access the copy of the item of `columns` by usage `options.column`. Thus one should be able to modify `searchoptions` by modifying of  `options.column.searchoptions`.