add support of mousewheel/DOMMouseScroll in setfrozenColumns
One can use now mouse wheel on frozen div to make vertical scrolling of the the grid. See [the answer](http://stackoverflow.com/a/39624573/315935) for more details. The offset for scrolling could be not perfect calculation. If one have the better way then one can use `mouseWheel` callback of `setFrozenColumns`. The below code uses the callback but with the same calculation algorithm line in the current code of free jqGrid:
```JavaScript
$grid.jqGrid("setFrozenColumns", {
    mouseWheel: function (e) {
        return e.type === "mousewheel" ?
            -e.originalEvent.wheelDelta / 10 :
            e.originalEvent.detail * 6;
    }
});
```