improve compatibility of edittype:"checkbox", which uses "change" event handler
The code of 4.13.3 and 4.13.4 triggers the "change" event after filling of select at the end of processing of `dataUrl`. It can produce compatibility problems if one uses `dataEvents: [ { type: "change", fn: function () {...} } ]` to monitor changes of the select made *by the user*. The "change" event could be interpreted by the event handler as the changes of the user, which will be wrong. To fix the problem in 4.13.3 and 4.13.4 one would need to move the code of the change handler from `dataEvents: [ { type: "change", fn: function () { ... } } ]` to `selectFilled: function (options) { $(options.elem).change(function () {...} }` as the workaround.

The current changes removes triggering of the "change" after filling of select at the end of processing of `dataUrl`. It makes unneeded to make any changes in the code, which worked in 4.13.2 or older and which used `dataEvents: [ { type: "change", fn: function () {...} } ]`.