Fix showHideCol for Microsoft Edge in case of working with empty grid
Microsoft Edge don't apply correctly changing if `display` style (adding/removing of `display: none`) on the empty grid. The fix makes the changes of the `display` style in 3 steps now on the empty grid (1. changing the `height` from `0` to `1px`; 2. changing of `display` style; 3. resetting the `height` to the original `0` value) as the workaround. Without the fix the grid header show/hide the correct columns, but the columns of the grid body stay hidden.