replace jQuery methods bind and unbind to on and off
The methods `bind` and `unbind` are the wrapper on `on` and `off` introduced in jQuery 1.7 (see [here](https://github.com/jquery/jquery/blob/1.7/src/event.js#L965-L970)). Starting with jQuery 3.0 the methods `bind` and `unbind` are declared as deprecated (see [here](http://api.jquery.com/category/deprecated/deprecated-3.0/)), but there are still exist in the code. We replace `bind` and `unbind` to `on` and `off` to make the code more relible for the next versions of jQuery.