fix min height of the btable to 1px to hold horizontal scrolling position during filtering to no rows
we add `height: 1px` to the btable. If there is at least one visible row then the CSS property will be not used, but if the body is empty and the width of the grid require creating horizontal scrollbar then the first row will has height 1px instead of 0 used typically and one will be able scroll the grid horizontaly.