Add support of `customUnaryOperations` option to specify the names of operation which required no data (like nu and nn)
Thanks @Tal (http://stackoverflow.com/users/4152866/tal) for pointing to existing problems (see [the question](http://stackoverflow.com/q/41441826/315935) for additional information). If the user chooses the predefined "is null" operation ("nu") or "is not null" operation ("nu") in the filter toolbar, the filtering will be start even with no data entered by user. In the same way, choosing the same operations in the Searching Dialog makes the input field, used for the data, readonly.

One can define new custom operations, which are **unary operations** in the same way like "is null" ("nu") or "is not null" ("nu"). One need to inform free jqGrid about the operations. New option `customUnaryOperations` allows to do that. The value of `customUnaryOperations` is the array with the names (strings) of the custom unary operations.