change the options of saveRowValidation callback of inline editing or the event jqGridInlineSaveRowValidation
The callback `saveRowValidation` or the event `jqGridInlineSaveRowValidation` have now the only parameter with the following list of properties:
- `mode` with the value `"edit"` or `"add"`
- `rowid` - the value of id attribute of the editing row
- `newData` - the object with modified data of the row
- `savedRow` - the object with the data of the row before editing
- `options` - the options of inline editing
- `iRow` - the index of the row from the top of the grid
- `tr` - DOM element, which represent the editing row

The callback `saveRowValidation` or the event `jqGridInlineSaveRowValidation` should return false or "stop" to report the error. It can either display the error message itself or to set the error message in `errorText` property of the input parameter. For example
```JavaScript
inlineEditing: {
 saveRowValidation: function (options) {
  if (options.newData.name.substr(0, 4) !== "test") {
   options.errorText = "The name have to start with \"test\"!";
   //alert("The name have to start with \"test\"!");
   return false;
  }
 }
}
```
sets the error message to `errorText` property and jqGrid displays the corresponding error message. One can remove the line whith the assignment of `options.errorText` property and uncomment the statement with `alert`.