Bug fix in processing custom buttons in info_dialog
Thanks @kosiaciok (https://github.com/kosiaciok) for reporting the bug. See [the issue](https://github.com/free-jqgrid/jqGrid/issues/284) for more details. The parameters of `onClick` is changed a little: `this` is now the DOM of the grid, the first parameter `e` is the Event of `click` event and the second parameter is jQuery wrapper on the current Info dialog.