Bug fix of filtering/searching for string properties, which could have null or undefined values
Thanks @Lijo (http://stackoverflow.com/users/696627/lijo) for reporting of the bug. See [the question](http://stackoverflow.com/q/41899798/315935) for more detailes.

Old versions of jqGrid convert every property of filterd items to string (if no `sorttype` is defined). As the result, the properties, which has `null` or `undefined` values, will be converted to the strings `"null"` or `"undefined"`. It could produce misunderstandings during filtering in case of the user search, for example, for `"n"`, `"u"` and so on. To fix the problem we replace the falsy values (`null` and `undefined`) to the empty string.