send additional parameter idOld to the server in case of id changes
If one have an editable column with `key: true` then both the old value and the new value of id will be send to the server. The default name of the new property with the old id value is `idOld`. It can be changed by specifying another value in `prmNames` parameter of jqGrid.