Add sorttype:"boolean" as the workaround for filtering by boolean false value
The commit https://github.com/free-jqgrid/jqGrid/commit/5d46f2241dcec1745966253c2276eecac6042e81 brokes filtering by boolean `false` value. The problems was: the boolen was interpreted as string and the expression `jQuery.jgrid.getAccessor(this,'closed') || ''` will replace `false` value to `''` instead of `'false'`. The later comparing of `''` with `'false'` will not detect equality.

I added `sorttype: "boolean"` as the workaround. It allows to produce the compare statement for boolean other as for strings. It allows to "normilize" any input data to Boolean `true` or `false` and then compare the results for equality.