Allow to specify classes with custom icons for TreeGrid nodes
The `icon` property of input data of TreeGrid already could specify the custom class of **leafs** of TreeGrid. The changes allows to specify custom icons of **nodes** of TreeGrid too. The nodes displays one from **two** icons depend on the state of the node (expanded or collapsed). The `icon` property can now specify the two icons, which need be comma separated. For example, one can use `icon: "ui-icon-folder-open,ui-icon-folder-collapsed"` in case of using jQuery UI icons.