New options of milti-sorting implemented: showSortOrder, multiSortOrder and formatSortOrder
The new option allows to display small numbers on the right side of sorting icons by usage `showSortOrder:true`. The option works only in combination with the option `multiSort:true`. The default behavior of the new feature is displaying 1-based number of sorting index. One can use `formatSortOrder` callback function to custom formatting of the index. For example, the following code
```JavaScript
formatSortOrder: function (options) {
    return "<i>" + (options.sortIndex + 1) + "</i>";
}
```
display the numbers as italic. The same changes one can do by changing CSS rules on `.ui-jqgrid-sort-order`. The default CSS settings are the following
```css
.ui-jqgrid-sortable > .ui-jqgrid-sort-order {
    position: relative; left: -.1em; top: 0; font-size: 75%; vertical-align: super;
}
```
One can consider to change `font-size` or `left` based on own preferences.

Another sinple example of the usage of `formatSortOrder` is below
```JavaScript
formatSortOrder: function (options) {
    // options has currently two propertis cm and sortIndex
    return String.fromCharCode("α".charCodeAt(0) + options.sortIndex);
}
```
The code force displaying of greek laters `α`, `β`, `γ`, `δ`, `ε`. instead of the standard arabic numbers `1`, `2`, `3`, `4`, `5`, ...

Another new option `multiSortOrder` can be used to change the **order** of sorting on click on the column headers. The old behavior is the following. Let us the grid is initallay sorted by columns A and B. The click on the column header C will sort the grid by A, B, C. The problem exists if the user want to *change* (to replace) the previous (or initial) sorting by A and B to *new* sorting by C. To do this in the standard behavior the use have to click twice on the column headers of A and B columns before clicking on the column header of C. Such behavior is not comfortable in some scenarios. One can use `multiSortOrder: "lastClickedFirstSorted"` to force the sorting order C, A, B instead of A, B, C after clicking on the header of C column in case of existing sorting order by A, B. If the user want to sort only ba C then the sorting by C, A, B gives mostly the same r very close results. We hold `multiSortOrder: "lastClickedLastSorted"` as default because of compatibility reasons, but we will consider to change the default value in the future. One more possibility to use the `multiSortOrder` option is specifying callback function which will be called only on click on new column headers. The code below is an example of such kind of usage. It do the same as the option `multiSortOrder: "lastClickedFirstSorted"`:
```JavaScript
multiSortOrder: function (sortNames, cm, sortDirs) {
    sortNames.unshift(sortNames[sortNames.length - 1]);
    sortNames.pop();
    return sortNames; // return new custom sort order
}
```