Add new multiselect options maxSortColumns, sortOrderPosition. Change default value od multiSortOrder and the options of multiSortOrder callback
The default value of `multiSortOrder` option is changes from old-syle `"lastClickedLastSorted"` to new-style `"lastClickedFirstSorted"`. New numeric option `maxSortColumns` with default value 3 allows to reduce the total number of column by which the grid could be sorted. It works only in case of usage `multiSortOrder: "lastClickedFirstSorted"` option. If the user clicks of the 4th column header and `maxSortColumns` is 3 then the column with the 4th sorting index will be removed and the grid will be sorted by 3 columns maximal. Such behavir simplify for the user removing of sorting by unneed columns.

New option `sortOrderPosition` with two allowed values `"afterSortIcons"` and `"beforeSortIcons"` provides more customization possibilities. Default value is `"afterSortIcons"`, but `sortOrderPosition: "beforeSortIcons"` could be interesing, for example, in case of combination with `sortIconsBeforeText: true` option.

The option `multiSortOrder` could be used as callback function additionally to the usage as `"lastClickedFirstSorted"` or `"lastClickedLastSorted"`. The callback has now only one `options` parameter with `sortNames`, `cm`, `sortDirs` and `removeSorting` properties. New property `removeSorting` is the function, which can be used by `multiSortOrder` callback to remove some column from sorting. As I mention above, the option `maxSortColumns` works only in case of usage `multiSortOrder: "lastClickedFirstSorted"`. If the callback want to implement close behavior it should be able to remove sorting order. In the case it can call `options.removeSorting` function with the element of `options.sortNames` array as parameter.