some changes to allow to use editoptions.value as function for edittype:"select" or/and formatter:"select"
The `editoptions` property be used by `formatter: "select"` if no `formatoptions.value` is defined. Old code of jqGrid allowed to use `editoptions.value` and `searchoptions.value` as function, but the function was not used by `formatter:"select"`. The modifications fixes the problem and all `editoptions.value`, `searchoptions.value` and `formatoptions.value` can be a function, which returns string or object in the well-known format of the `value` property. It's important to mention, that `value` as function can be used only by `edittype:"select"` or/and `formatter:"select"` and not for all other possible values of `edittype` and `formatter`.