add 3 new option: sortingDuringEditing, pagingDuringEditing, reloadingDuringEditing with values "prevent", "cancel" or "save"
The default behavior in previous versions of jqGrid: preventing sorting if the grid is in inline or in cell editing. Including new option `sortingDuringEditing: "cancel"` or `sortingDuringEditing: "save"` will alows sorting. The currently editing data will be discarded or saved depend on the value of `sortingDuringEditing` option. The same problem exists in case of paging or reloading the grid. The oprions `pagingDuringEditing`, `reloadingDuringEditing` helps to specify the desired behavior.